requirements.txt
acme==1.4.0
asgiref==3.2.7
beautifulsoup4==4.9.0
boto3==1.9.179
botocore==1.12.179
bs4==0.0.1
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
ConfigArgParse==1.2.3
configobj==5.0.6
cryptography==2.9.2
distro==1.5.0
dj-database-url==0.5.0
Django==2.2.13
django-crispy-forms==1.7.2
django-endless-pagination==2.0
django-environ==0.4.5
django-extensions==2.1.6
django-heroku==0.3.1
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-storages==1.7.1
django-unused-media==0.1.13
docutils==0.14
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
jdcal==1.4
jmespath==0.9.4
josepy==1.3.0
jsonfield==2.0.2
notifications==0.3.2
openpyxl==2.6.1
parsedatetime==2.5
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
pyOpenSSL==19.1.0
pyRFC3339==1.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.9.1
s3transfer==0.2.1
six==1.12.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.24.2
whitenoise==4.1.3
django-imagefield==0.12.4